Move Definition::{Hash,Array}=>Types::{Hash,Array}
